requirements/requirements.txt
appconfigpy>=0.3.0,<1.0.0
colorama>=0.4.1,<1.0.0
DateTimeRange>=0.5.3,<1.0.0
enum34;python_version<"3.4"
Logbook>=1.1.0,<2.0.0
mbstrdecoder[all]>=0.6.2,<1.0.0
msgfy>=0.0.5,<0.1.0
pathvalidate>=0.23.1,<1.0.0
PyGithub>=1.43.4,<2.0.0
pytablewriter>=0.41.1,<1.0.0
requests>=2.21.0,<3.0.0
simplejson>=3.16.0,<4.0.0
subprocrunner>=0.16.0,<1.0.0
tqdm>=4.29.1,<5.0.0
